feat: refactor structure for SSL
