+++ src/pages/ssl/pages/create/components/Step3/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Form, Button } from 'antd';
-import { create as createSSL } from '@/services/ssl';
+import { create as createSSL } from '@/pages/ssl/services/request';
 import CertificateForm from '../CertificateForm';
 import { StepProps } from '../..';
