+++ src/pages/ssl/pages/list/index.tsx
@@ -4,7 +4,7 @@ import ProTable, { ProColumns, ActionType } from '@ant-design/pro-table';
 import { Button, Modal, notification, Switch } from 'antd';
 import { history, useIntl } from 'umi';
 import { PlusOutlined } from '@ant-design/icons';
-import { fetchList as fetchSSLList, remove as removeSSL } from '@/services/ssl';
+import { fetchList as fetchSSLList, remove as removeSSL } from '@/pages/ssl/services/request';
 import { SSL } from '@/models/ssl';
 import { ListItem } from '@/transforms/global';
